Bug fix in forward and bi-directional lstm in handling bias
The converter was not setting the value of R bias correctly during
onnx conversion.

We simply set the R biases to zero since the current conversion code
is already setting W bias to the Keras' bias value. We expect the sum
of the biases Wb + Rb to be the same as Keras' bias since Keras uses
a single bias combining both.

Changes to be committed:

	modified:   keras2onnx/ke2onnx/bidirectional.py
	modified:   keras2onnx/ke2onnx/lstm.py
	modified:   tests/test_layers.py